refactor: dto분리
