+++ HW3SpringBoard/src/main/java/com/seol/HW3SpringBoard/dto/PostDTO.java
@@ -1,4 +1,4 @@
-package com.seol.HW3SpringBoard.domain;
+package com.seol.HW3SpringBoard.dto;
 
 import lombok.Getter;
 import lombok.Setter;
